[BE] Added authorization of registration, login, password change, schedule, Tasks, Account balance

+++ backend/SportsCenterBackend/SportsCenter.Application/Employees/Commands/RemoveTask/RemoveTaskHandler.cs
@@ -23,18 +23,21 @@ namespace SportsCenter.Application.Employees.Commands.RemoveTask
         }
         public async Task<Unit> Handle(RemoveTask request, CancellationToken cancellationToken)
         {
-            //pytanie czy zwykly pracownik moze usuwac wszystkie zadania
-            //czy np tylko te ktore maja jako IdZlecajacego jego Id
-            //czyli tylko te ktore sam sobie przydzielil -> jesli tak to wziac jego id z sesji  i porownac z IdZlecajacego
-
-
-            //obecna wersja zezwala na usuwanie przez pracownika kazdego zadania z listy
+            if (!int.TryParse(_httpContextAccessor.HttpContext?.User.FindFirst("userId")?.Value, out int userId))
+            {
+                throw new UnauthorizedAccessException("No permission to remove tasks.");
+            }
 
             var task = await _employeeRepository.GetTaskByIdAsync(request.TaskId, cancellationToken);
 
             if (task == null)
             {
                 throw new TaskNotFoundException(request.TaskId);
+            }
+
+            if (task.PracownikId != userId)
+            {
+                throw new UnauthorizedAccessException("You can only remove your own tasks.");
             }
 
             await _employeeRepository.RemoveTaskAsync(task, cancellationToken);
